VIIRS Climate Raw Data Record (C-RDR) example. (#7)
This is an example of NOAA Level 1b satellite swath data from the Suomi-NPP VIIRS instrument stored in netCDF-4. It uses an early version of ACDD (pre-ESIP) and metadata derived from the NOAA Suomi-NPP Raw Data Record (RDR) and Sensor Data Record (SDR) HDF5 file specifications. This data set is intended for use by reprocessors. It attempts to encode much of the information about the measurements from the VIIRS instrument documentation to make it highly self-describing, and to make all measurements much easier to access in comparison to using the VIIRS RDR data set, which is composed of cleaned up CCSDS data packets.